Algoritmos de consenso ya funcionan

+++ agregador/src/main/java/aplicacion/domain/algoritmos/AlgoritmoConsensoIrrestricto.java
@@ -11,7 +11,7 @@ import java.util.Map;
 @DiscriminatorValue("irrestricto")
 public class AlgoritmoConsensoIrrestricto extends AlgoritmoConsenso {
     @Override
-    public List<Hecho> curarHechos(Map<Hecho, Integer> cantidadPorHecho, Integer totalFuentes) {
+    public List<Hecho> curarHechos(Map<Hecho, Long> cantidadPorHecho, Long totalFuentes) {
         return new ArrayList<>(cantidadPorHecho.keySet());
     }
 }
