+++ agregador/src/main/java/aplicacion/domain/algoritmos/AlgoritmoConsensoMayoriaSimple.java
@@ -9,7 +9,7 @@ import java.util.*;
 @DiscriminatorValue("mayoriaSimple")
 public class AlgoritmoConsensoMayoriaSimple extends AlgoritmoConsenso {
     @Override
-    public List<Hecho> curarHechos(Map<Hecho, Integer> cantidadPorHecho, Integer totalFuentes) {
+    public List<Hecho> curarHechos(Map<Hecho, Long> cantidadPorHecho, Long totalFuentes) {
         int umbral = (int) Math.ceil(totalFuentes / 2.0);
         return cantidadPorHecho.entrySet().stream()
                 .filter(e -> e.getValue() >= umbral)
